Move BranchFolding.h to include/
BranchFolder cleanup is required for the new jump table-related feature.
In order for our pass to access the BranchFolding.h header file, it has
been moved to the include/ directory.
